cambio de int a double para que soporte cantidades mayores

+++ src/Boletin2_9.java
@@ -17,10 +17,10 @@ public class Boletin2_9 {
      * @param args the command line arguments
      */
     public static void main(String[] args) {
-        int b100=Integer.parseInt(JOptionPane.showInputDialog("¿Cuantos billetes hay de 100€?"));
-        int b20=Integer.parseInt(JOptionPane.showInputDialog("¿Cuantos billetes hay de 20€?"));
-        int b5=Integer.parseInt(JOptionPane.showInputDialog("¿Cuantos billetes hay de 5€?"));
-        int b1=Integer.parseInt(JOptionPane.showInputDialog("¿Cuantas monedas hay de 1€?"));
+        double b100=Double.parseDouble(JOptionPane.showInputDialog("¿Cuantos billetes hay de 100€?"));
+        double b20=Double.parseDouble(JOptionPane.showInputDialog("¿Cuantos billetes hay de 20€?"));
+        double b5=Double.parseDouble(JOptionPane.showInputDialog("¿Cuantos billetes hay de 5€?"));
+        double b1=Double.parseDouble(JOptionPane.showInputDialog("¿Cuantas monedas hay de 1€?"));
         JOptionPane.showMessageDialog(null, "En total hay "+(b100*100+b20*20+b5*5+b1)+"€");
     }
